refactor: collapse capability switches and split autorestart strategy
