+++ internal/strategy/routes.go
@@ -7,26 +7,6 @@ import (
 	"github.com/pinchtab/pinchtab/internal/routes"
 )
 
-// capabilityEnabled maps a routes.Capability to the orchestrator check.
-func capabilityEnabled(orch *orchestrator.Orchestrator, cap routes.Capability) bool {
-	switch cap {
-	case routes.CapEvaluate:
-		return orch.AllowsEvaluate()
-	case routes.CapMacro:
-		return orch.AllowsMacro()
-	case routes.CapScreencast:
-		return orch.AllowsScreencast()
-	case routes.CapDownload:
-		return orch.AllowsDownload()
-	case routes.CapUpload:
-		return orch.AllowsUpload()
-	case routes.CapStateExport:
-		return orch.AllowsStateExport()
-	default:
-		return false
-	}
-}
-
 // capabilitySetting returns the config key and disabled code for a capability.
 func capabilitySetting(cap routes.Capability) (feature, setting, code string) {
 	switch cap {
@@ -59,7 +39,7 @@ func RegisterShorthandRoutes(mux *http.ServeMux, orch *orchestrator.Orchestrator
 	}
 
 	for cap, eps := range routes.CapabilityEndpoints() {
-		enabled := capabilityEnabled(orch, cap)
+		enabled := orch.Allows(cap)
 		feature, setting, code := capabilitySetting(cap)
 		for _, ep := range eps {
 			RegisterCapabilityRoute(mux, ep.Route(), enabled, feature, setting, code, handler)
